Finalized committee group page.
Added support for a committee picture in the group page, but since there is no good structure yet I have not implemented it yet. Or you could hard code the image links but that seems dumb so I have commented it. If you uncomment it you should be able to add a image which is displayed to the right of the group info.

Furthermore the committees page has been cleaned up and now links to the correct group which displays the basic information of the group in the following order:

fullName
description
creation date
email
members

The only TODO left is to remove the css that is no longer used from the old committee page, but I was not sure what was safe to delete and what was FUBAR.